🚚 fetchAPI index로 변경
