--- apis/fetchAPI.ts
@@ -1,66 +1,0 @@
-import { convertObjToURLSearchParams } from '@/utils/convert';
-
-const BASE_URL = 'http://cse-dev-waffle.bacchus.io';
-
-const paramsToString = (params: URLSearchParams) => (params.size ? `?${params}` : '');
-
-export const getRequest = async <T = unknown>(
-  url: string,
-  params: object = {},
-  headers: HeadersInit = {},
-) => {
-  const urlSearchParams = convertObjToURLSearchParams(params);
-  const fetchUrl = `${BASE_URL}${url}${paramsToString(urlSearchParams)}`;
-  try {
-    const response = await fetch(fetchUrl, { method: 'GET', headers });
-    const responseData = await response.json();
-    return responseData as T;
-  } catch (error) {
-    console.log('error on get request');
-  }
-};
-
-export const postRequest = async <T = unknown>(
-  url: string,
-  data: object,
-  headers: HeadersInit = {},
-) => {
-  const fetchUrl = `${BASE_URL}${url}`;
-  try {
-    const response = await fetch(fetchUrl, { method: 'POST', headers, body: JSON.stringify(data) });
-    const responseData = await response.json();
-    return responseData as T;
-  } catch (error) {
-    console.log('error on post request');
-  }
-};
-
-export const patchRequest = async <T = unknown>(
-  url: string,
-  data: object,
-  headers: HeadersInit = {},
-) => {
-  const fetchUrl = `${BASE_URL}${url}`;
-  try {
-    const response = await fetch(fetchUrl, {
-      method: 'PATCH',
-      headers,
-      body: JSON.stringify(data),
-    });
-    const responseData = await response.json();
-    return responseData as T;
-  } catch (error) {
-    console.log('error on patch request');
-  }
-};
-
-export const deleteRequest = async <T = unknown>(url: string, headers: HeadersInit = {}) => {
-  const fetchUrl = `${BASE_URL}${url}`;
-  try {
-    const response = await fetch(fetchUrl, { method: 'DELETE', headers });
-    const responseData = await response.json();
-    return responseData as T;
-  } catch (error) {
-    console.log('error on delete request');
-  }
-};
